+++ R/pas_filterNear.R
@@ -3,14 +3,14 @@
 #' @title Find PurpleAir sensors within radial distance
 #' 
 #' @param pas PurpleAir \emph{pas} object.
-#' @param latitude a Target latitude. 
 #' @param longitude a Target longitude.
+#' @param latitude a Target latitude. 
 #' @param radius Distance from target with unit (i.e "15 km").
 #' 
 #' @description Filter for Purple Air sensors within a specified distance from 
 #' specified target coordinates. 
 #' 
-#' @details \code{radius} Should be a numeric string with provided metric unit 
+#' @details \code{radius} Should be a numeric string with a metric unit 
 #' separated by a space, such as \code{"250 m"}.
 #'
 #' @return A subset of the given \emph{pas} object.
@@ -23,8 +23,8 @@
 #' pas <- example_pas
 #' diamond_bar <-
 #'   pas %>%
-#'   pas_filterNear(latitude = 34.001667, 
-#'                  longitude = -117.820833,
+#'   pas_filterNear(longitude = -117.820833,
+#'                  latitude = 34.001667, 
 #'                  radius = "20 km")
 #'                  
 #'\dontrun{
@@ -33,23 +33,23 @@
 
 pas_filterNear <- function(
   pas = NULL,
-  latitude = NULL, 
   longitude = NULL, 
+  latitude = NULL, 
   radius = "1 km"
 ) {
   
   # ----- Validate parameters --------------------------------------------------
   
   MazamaCoreUtils::stopIfNull(pas)
+  MazamaCoreUtils::stopIfNull(longitude)
+  MazamaCoreUtils::stopIfNull(latitude)
+  MazamaCoreUtils::stopIfNull(radius)
   
   if ( !pas_isPas(pas) )
     stop("Required parameter 'pas' is not a valid 'pa_synoptic' object.")
   
   if ( pas_isEmpty(pas) )
     stop("Required parameter 'pas' has no data.") 
-  
-  if ( is.null(latitude) || is.null(longitude) )
-    stop("Required target coordinate(s) is missing") 
   
   if ( !stringr::str_ends(radius, "[ km]") )
     stop("Radius requires a unit and format (i.e '1 m' or '1 km')")