+++ vignettes/articles/outlier_detection.R
@@ -1,4 +1,0 @@
-## ----setup, include=FALSE, message=FALSE---------------------------------
-knitr::opts_chunk$set(fig.width=7, fig.height=5)
-library(AirSensor)
-